test(logging): cover schemas with zod
